Begin new project Etch-a-Sketch
This commit includes three new files, one .html, .css. and .js. The
necessary functions to create a 16x16 grid are included in the .js file.

The issue that needs to be overcome is to make the grid flexible and
able to grow and shrink with the window without changing its
dimensions.