Add functionality to operations and display
This commit adds functionality to the operation buttons. The
functionality of only the '+' was tested however, errors were
thrown for some other buttons. The new changes allow the user to
input a number, choose an operation, and click a number. If they
choose another operation, the display is updated to display the
result of the previous operation.