Change functionality and fix bugs
This commit changed the functionality of the calculator from the initial
concept in the main branch. The user now must click the 'equals' button
in order to process an operation. This bypasses complications with
trying to determine in the code whether they are trying to change the
operation they selected, or chaining them.

Event listeners have been added in the calculator.js file, and the
'onclick' properties removed from index.html. Classes from the 'equals'
and 'clear' buttons have been removed to differentiate their functionality
from the other buttons.